// src/collection/domain/models/CollectionType.ts
export enum CollectionType {
  RESEARCHERS = 'RESEARCHERS',
  RESEARCH_PROJECTS = 'RESEARCH_PROJECTS',
  JOURNALS = 'JOURNALS',
  ORGANIZATIONS_INSTITUTIONS = 'ORGANIZATIONS_INSTITUTIONS',
  TEACHING_COURSES = 'TEACHING_COURSES',
  UNCATEGORIZED = 'UNCATEGORIZED',
  LABORATORY = 'LABORATORY',
  RESEARCH_GROUP = 'RESEARCH_GROUP',
  DEPARTMENT = 'DEPARTMENT'
}